lp-522 Reverse strange refactor
