+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/dto/DataDto.java
@@ -14,8 +14,8 @@ import uk.gov.companieshouse.limitedpartnershipsapi.model.dto.validator.NameSize
 import uk.gov.companieshouse.limitedpartnershipsapi.model.dto.validator.ValidEnum;
 
 import static com.fasterxml.jackson.annotation.JsonInclude.Include.NON_NULL;
-import static uk.gov.companieshouse.limitedpartnershipsapi.model.dto.validator.utils.Constants.LONG_MAX_SIZE;
-import static uk.gov.companieshouse.limitedpartnershipsapi.model.dto.validator.utils.Constants.MIN_SIZE;
+import static uk.gov.companieshouse.limitedpartnershipsapi.utils.Constants.LONG_MAX_SIZE;
+import static uk.gov.companieshouse.limitedpartnershipsapi.utils.Constants.MIN_SIZE;
 
 @NameSize
 public class DataDto {
