+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/model/dto/validator/NameSize.java
@@ -8,7 +8,7 @@ import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.lang.annotation.Target;
 
-import static uk.gov.companieshouse.limitedpartnershipsapi.model.dto.validator.utils.Constants.LONG_MAX_SIZE;
+import static uk.gov.companieshouse.limitedpartnershipsapi.utils.Constants.LONG_MAX_SIZE;
 
 @Constraint(validatedBy = NameSizeValidator.class)
 @Target(ElementType.TYPE)
