+++ src/main/java/uk/gov/companieshouse/limitedpartnershipsapi/utils/ApiLogger.java
@@ -1,4 +1,4 @@
-package uk.gov.companieshouse.limitedpartnershipsapi.model.dto.validator.utils;
+package uk.gov.companieshouse.limitedpartnershipsapi.utils;
 
 import uk.gov.companieshouse.limitedpartnershipsapi.LimitedPartnershipsApiApplication;
 import uk.gov.companieshouse.logging.Logger;
